impute: filling in missing data

This impute function is used to fill in the missing values of a data frame using a specified method from the following three: knn, missForest or mice. If the method input is not one of the three previously listed, then the function will throw an error.

impute(data, method = c("knn", "missForest", "mice"))

Arguments

data

data frame with missing values

method

method to be applied to impute the missing values

Value

a new data frame where the missing value are estimated based on the specified method

Details

The impute function is a wrapper for the following functions: knn function, missForest function, mice function, and merge_imputation function.

Examples

# airquality is a data frame from base R

new_df <- impute(airquality, "knn")
new_df <- impute(airquality, "missForest")
new_df <- impute(airquality, "mice")

# if the specified method is not among the listed three, an error will be reported
new_df <- impute(airquality, "randomforest")
#> Error in match.arg(method): 'arg' should be one of “knn”, “missForest”, “mice”